Fix diffs/rebase

+++ contracts/Comet.sol
@@ -747,7 +747,7 @@ contract Comet is CometMainInterface {
     /**
      * @dev Divide a common price quantity by a price, returning a `toScale` quantity
      */
-    function divPrice(uint n, uint price, uint toScale) internal pure returns (uint) {
+    function divPrice(uint n, uint price, uint64 toScale) internal pure returns (uint) {
         return n * toScale / price;
     }
 
@@ -1138,7 +1138,7 @@ contract Comet is CometMainInterface {
     }
 
     /**
-     * @dev Transfer user's collateral and debt to the protocol itself
+     * @dev Transfer user's collateral and debt to the protocol itself.
      */
     function absorbInternal(address absorber, address account) internal {
         if (!isLiquidatable(account)) revert NotLiquidatable();
